refactor: 일일 통계 이벤트 구조 리팩토링
- internal event 패키지로 정의
- DailyStampStatisticsEvent로 이벤트 객체 이름 변경

Resolved: #44
