+++ src/main/java/shop/genieus/study/domains/statistics/application/DailyStatisticsService.java
@@ -12,11 +12,11 @@ import shop.genieus.study.commons.provider.AttendanceProvider;
 import shop.genieus.study.commons.provider.DateTimeProvider;
 import shop.genieus.study.commons.provider.StampHistoryProvider;
 import shop.genieus.study.commons.provider.UserProvider;
-import shop.genieus.study.commons.provider.dto.AttendanceInfo;
-import shop.genieus.study.commons.provider.dto.StampHistoryInfo;
-import shop.genieus.study.commons.provider.dto.UserInfo;
+import shop.genieus.study.commons.provider.model.AttendanceInfo;
+import shop.genieus.study.commons.provider.model.StampHistoryInfo;
+import shop.genieus.study.commons.provider.model.UserInfo;
 import shop.genieus.study.domains.statistics.application.assembler.DailyStatisticsAssembler;
-import shop.genieus.study.domains.statistics.domain.event.DailyStatisticsEvent;
+import shop.genieus.study.domains.statistics.application.event.internal.model.DailyStampStatisticsEvent;
 import shop.genieus.study.domains.statistics.domain.vo.UserStatistics;
 
 @Slf4j
@@ -62,7 +62,7 @@ public class DailyStatisticsService {
           assembler.assembleUserStatistics(
               userIds, userNicknameMap, attendanceInfos, stampHistoryInfos);
 
-      DailyStatisticsEvent event = new DailyStatisticsEvent(yesterday, userStatistics);
+      DailyStampStatisticsEvent event = new DailyStampStatisticsEvent(yesterday, userStatistics);
       eventPublisher.publishEvent(event);
 
       log.info("일일 통계 알림 발송 완료: {} ({} users)", yesterday, userStatistics.size());
